backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
pypdf
python-multipart
python-dotenv
transformers
torch
sentence-transformers
